Write a new go-list function for package imports
This commit adds a function that might just make my professional life as a go
programmer bearable.

go-mode comes with two baked-in functions for listing imports. One is pure
elisp, and is slow and inaccurate. The other wraps `go list -e all`, and is
accurate, but unusably slow. This is a third option. It relies on projectile,
which means it isn't really portable. But: it restricts imports to *only* the
standard lib, internal-to-a-project packages, and the vendor dir of the current
project. It's fast enough to use *and* it's accurate.

WINNING.